Move exponential_backoff test to test_utils.py
