+++ tests/test_utils.py
@@ -1,6 +1,8 @@
 import asyncio
+import itertools
 import os
 import random
+from typing import List
 
 import pytest
 
@@ -9,6 +11,7 @@ from httpx._utils import (
     ElapsedTimer,
     NetRCInfo,
     URLPattern,
+    exponential_backoff,
     get_ca_bundle_from_env,
     get_environment_proxies,
     guess_json_utf,
@@ -270,3 +273,16 @@ def test_pattern_priority():
         URLPattern("http://"),
         URLPattern("all://"),
     ]
+
+
+@pytest.mark.parametrize(
+    "factor, expected",
+    [
+        (0.1, [0, 0.1, 0.2, 0.4, 0.8]),
+        (0.2, [0, 0.2, 0.4, 0.8, 1.6]),
+        (0.5, [0, 0.5, 1.0, 2.0, 4.0]),
+    ],
+)
+def test_exponential_backoff(factor: float, expected: List[int]) -> None:
+    delays = list(itertools.islice(exponential_backoff(factor), 5))
+    assert delays == expected
